Ajustando coerência com banco de dados

+++ src/main/java/br/com/fiap/spacecrop/dto/response/FazendaResponseDTO.java
@@ -14,6 +14,6 @@ public class FazendaResponseDTO {
     private String nome;
     private String cidade;
     private String estado;
-    private Double areaHectares;
+    private Float areaHectares;
     private Long usuarioId;
 }
